requirements.txt
# Core dependencies
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
structlog==24.1.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
